+++ docs/version-history.md
@@ -1,5 +1,11 @@
 # Version History
 
+### Version 1.61 (Aug 2, 2025)
+- Added A/B switching comparison function for pipelines
+- Added Crossfeed Filter effect
+- Added new version update check function
+- Various minor improvements
+
 ### Version 1.60 (Jul 27, 2025)
 - Added Comb Filter effect
 - Various minor improvements